refactor: split aces_conformance/conformance.py into subdomain modules
